gen results v2
made packages for many of the functions. Added a new script for generating hpo EWCE results using all these. It seems to work better than the previous one. I just need to add the bits for downloading the required data from the HPO, so it doesnt need to be downloaded manually